requirements.txt
mediapipe==0.10.14
numpy==1.22.0
onnx==1.17.0
onnxruntime_gpu==1.22.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
